Add key_id and fingerprinting for public key packets
This is implemented for version 4 keys, which should be the only ones
anyone needs at this point. The algorithm described in RFC 4880 is
implemented and seems to be working as expected, and test cases have
been added. The key_id attribute is meant to be directly comparable with
the key_id from signature packets.

Additional changes here include parsing DSA and ElGamal key components;
the original 'mod' and 'exp' attributes from RSA keys have been renamed
to 'modulus' and 'exponent' for clarity.

Signed-off-by: Dan McGee <dan@archlinux.org>