Always represent key_id and fingerprint as byte strings
Treating key_id as an integer led to several headaches, especially when
it came to leading zeroes and trying to determine whether you have a
160-bit, 64-bit, or 32-bit identifier. Make it a string on all relevant
packets so it can be directly compared and looked at in an expected
fashion.

Also ensure fingerprint is treated as a bytestring everywhere, not a
Unicode string.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>