Deprecate datetime attribute on packets; add expiration_date
The datetime attribute works if there is only one datetime field in
play, but signatures can have both a creation and expiration date. Use
the same convention we use elsewhere, moving the raw epoch value to a
'raw_' prefixed attribute and using the 'creation_time' and
'expiration_time' attributes for the Python datetime objects.

Note the special handling of expiration_time- we use Python timedelta
objects to translate it to an actual datetime object, since the raw
value is just a seconds offset from the creation date.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>