Add get_int_bytes helper function
This ensures we do proper conversion of all integers back to their
big-endian byte string representations. We were blowing up on 65537,
which is '010001' in hex, which resulted in the following traceback:

    File "/home/muelli/git/python-pgpdump/pgpdump/packet.py", line 331, in
    get_bytes = lambda x: binascii.unhexlify(('%X' % x).encode('ascii'))
    TypeError: Odd-length string

Ensure the string is always padded to an even length and add tests for
the number that was breaking.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>