packet: Added a modulus_bitlen field to indicate the size of the RSA key
I was hesitating to add this, because it's a calculation that is not
necessarily needed by the user of pgpdump. But then again, that is the
case for nearly every data computed by pgpdump. It'd probably be smart
to make all fields needing calculation a lazy Property() and update the
fields after calculation.