Implement parsing of mode 1002 secret key packets
This is a GnuPG extension when the secret key data is stored on a smart
card or in a hardware security module. We blew up before when
encountering this unexpected type; now we will parse and continue on
without problems.

Based on work by Ondrej Mikle and reading RFC 4880 and taking a quick
glance at the GnuPG source.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>